+++ packages/api-headless-cms/src/content/plugins/graphqlFields/object.ts
@@ -3,6 +3,7 @@ import { CmsModelField, CmsModelFieldToGraphQLPlugin } from "~/types";
 import { renderField } from "~/content/plugins/utils/renderFields";
 import { renderInputField } from "~/content/plugins/utils/renderInputFields";
 import { createManageTypeName, createTypeName } from "~/content/plugins/utils/createTypeName";
+import { attachRequiredFieldValue } from "~/content/plugins/graphqlFields/requiredField";
 
 interface TypeFromFieldResponse {
     fieldType: string;
@@ -78,10 +79,7 @@ const plugin: CmsModelFieldToGraphQLPlugin = {
             });
 
             if (!result) {
-                return {
-                    fields: "",
-                    typeDefs: ""
-                };
+                return null;
             }
             const { fieldType, typeDefs } = result;
 
@@ -118,10 +116,7 @@ const plugin: CmsModelFieldToGraphQLPlugin = {
             });
 
             if (!result) {
-                return {
-                    fields: "",
-                    typeDefs: ""
-                };
+                return null;
             }
             const { fieldType, typeDefs } = result;
 
@@ -139,15 +134,15 @@ const plugin: CmsModelFieldToGraphQLPlugin = {
                 fieldTypePlugins
             });
             if (!result) {
-                return {
-                    fields: "",
-                    typeDefs: ""
-                };
+                return null;
             }
             const { fieldType, typeDefs } = result;
 
             return {
-                fields: `${field.fieldId}: ${field.multipleValues ? `[${fieldType}!]` : fieldType}`,
+                fields: attachRequiredFieldValue(
+                    `${field.fieldId}: ${field.multipleValues ? `[${fieldType}!]` : fieldType}`,
+                    field
+                ),
                 typeDefs
             };
         },