+++ packages/api-page-builder-so-ddb-es/src/upgrades/v5.15.0/category.ts
@@ -14,6 +14,7 @@ type DbRecord<T> = T & {
 export const upgradeCategories = async (context: PbContext) => {
     const tenant: string = context.tenancy.getCurrentTenant().id;
     const localeCodes: string[] = await context.i18n.getLocales().map(locale => locale.code);
+    // @ts-ignore
     const categoryStorageOperations = context.pageBuilder.categories
         .storageOperations as CategoryStorageOperationsDdbEs;
     if (categoryStorageOperations instanceof CategoryStorageOperationsDdbEs === false) {